チャート部分をコンポーネント化
