+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as CheckboxList } from './checkboxList';
 export { default as PopulationCategorySelect } from './populationCategorySelect';
+export { default as LineChartPopulation } from './lineChart';
