+++ src/features/Index/index.tsx
@@ -1,18 +1,15 @@
 import { Stack } from '@chakra-ui/react';
-import { LineChart, Line, CartesianGrid, XAxis, YAxis, Tooltip, Label, Legend } from 'recharts';
 import { useEffect, useState } from 'react';
 
 import { Prefecture, SelectedPrefectureData, YearlyPopulationData } from '@/types/index';
 import { getPopulationData, getPrefectures } from '@/functions/getPeopleData';
-import { CheckboxList, PopulationCategorySelect } from '@/components';
+import { CheckboxList, PopulationCategorySelect, LineChartPopulation } from '@/components';
 import Layout from '@/layout/layout';
 
 const IndexPage = (): JSX.Element => {
   const [prefectures, setPrefectures] = useState<Prefecture[]>([]);
   const [selectedPrefecturesData, setSelectedPrefecturesData] = useState<SelectedPrefectureData>({});
   const [selectedPopulationCategory, setSelectedPopulationCategory] = useState(0);
-
-  const COLOR_LIST = ['#FF0000', '#0000FF', '#00FF00', '#000000', '#800080', '#FFFF00'];
 
   useEffect(() => {
     const fetchPrefectures = async () => {
@@ -88,25 +85,7 @@ const IndexPage = (): JSX.Element => {
           selectedPrefecturesData={selectedPrefecturesData}
           handleCheckboxChange={handleCheckboxChange}
         />
-        <LineChart
-          width={800}
-          height={400}
-          data={generateChartData()}
-          margin={{ top: 10, right: 10, left: 80, bottom: 80 }}
-        >
-          {Object.entries(selectedPrefecturesData).map(([prefCode, { name }], index) => (
-            <Line type="monotone" dataKey={name} stroke={COLOR_LIST[index % COLOR_LIST.length]} key={prefCode} />
-          ))}
-          <CartesianGrid stroke="#ccc" />
-          <XAxis dataKey="year">
-            <Label value="年" offset={-20} position="insideBottom" />
-          </XAxis>
-          <YAxis>
-            <Label value="人口" angle={-90} position="insideLeft" offset={-50} />
-          </YAxis>
-          <Tooltip />
-          <Legend verticalAlign="top" wrapperStyle={{ lineHeight: '40px' }} />
-        </LineChart>
+        <LineChartPopulation selectedPrefecturesData={selectedPrefecturesData} generateChartData={generateChartData} />
       </Stack>
     </Layout>
   );
